# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: VertexLearn AI API
  version: 0.7.0
  description: Phase 7 surface — auth, course foundation, learning core, AI tutor/RAG (course-isolated), summaries, AI quiz drafts, flashcards, study plans, mastery, recommendations, admin workspace, course approvals, discussions, announcements, notifications, email abstraction, S3/MinIO storage (uploads, signed URLs), catalog filters/reviews, payments/revenue, Redis caching (catalog/course/leaderboard, x-cache), rate limits (auth 10/min/IP, AI 20/min/user, global 100/min), gamification leaderboard, health/ready aliases.
servers:
  - url: http://localhost:4000
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
                role: { type: string, enum: [student, instructor] }
      responses:
        '201': { description: Created }
        '400': { description: Validation error }
        '409': { description: Email taken or admin self-registration }
  /api/v1/auth/login:
    post:
      summary: Login, returns access + refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Authenticated }
        '401': { description: Invalid credentials }
  /api/v1/auth/refresh:
    post:
      summary: Rotate refresh token (single-use)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: New token pair }
        '401': { description: Invalid or expired refresh token }
  /api/v1/auth/logout:
    post:
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '204': { description: Revoked }
  /api/v1/auth/me:
    get:
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user }
        '401': { description: Unauthorized }
...
  /api/v1/courses/{id}:
    get:
      summary: Public course detail with modules + lectures
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Course detail }
        '404': { description: Not found }
    put:
      summary: Update own course (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
        '403': { description: Not the owner }
        '404': { description: Not found }
...
  /api/v1/courses/{courseId}/announcements:
    get:
      summary: List announcements (enrolled student, owner instructor, or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Announcements }
        '403': { description: Enrollment required }
        '404': { description: Course not found }
    post:
      summary: Publish an announcement (own course instructor or admin; notifies enrolled students)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Announcement created }
        '403': { description: Not the course owner }
        '404': { description: Course not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT